+++ CHANGELOG.md
@@ -36,6 +36,7 @@ configured prefix accordingly.
 - Added attribute `cogs.voice.MediaItem.user_id` to replace `.user`
 - Added method `console.LydianConsole._debug_evaluate_in_context()`
 - Added function `util.cog_commands()`
+- Added function `util.getclass()`
 - Added function `util.get_text_sections()`
 - Added function `util.mention()`